Move compare utils to utils/sort.ts in 'client' package
