+++ packages/client/src/__testData__/attendance.ts
@@ -2,9 +2,10 @@ import { CustomerWithAttendance } from "@/types/components";
 
 import { AttendanceCardProps } from "@/components/atoms/AttendanceCard";
 
+import { comparePeriods } from "@/utils/sort";
+
 import { baseSlot } from "@/__testData__/slots";
 import { gus, walt, saul, jian } from "@/__testData__/customers";
-import { comparePeriods } from "@/utils/helpers";
 
 export const allCustomers = [gus, walt, saul, jian];
