+++ packages/client/src/store/selectors/customers.ts
@@ -1,9 +1,10 @@
+import { DateTime } from "luxon";
+
 import { Customer, CustomersByBirthday } from "@eisbuk/shared";
 
 import { LocalStore } from "@/types/store";
 
-import { compareCustomers } from "@/utils/customers";
-import { DateTime } from "luxon";
+import { compareCustomerNames } from "@/utils/sort";
 
 /**
  * Get a record of all the customers for current organization from firebase store
@@ -25,7 +26,7 @@ export const getCustomersList =
   (sort?: boolean) =>
   (state: LocalStore): Customer[] => {
     const customerList = Object.values(getCustomersRecord(state) || {});
-    return sort ? customerList.sort(compareCustomers) : customerList;
+    return sort ? customerList.sort(compareCustomerNames) : customerList;
   };
 
 /**
